bot-files/requirements.txt
opencv-python
pygame
requests
openai
ytmusicapi
python-vlc
SpeechRecognition
azure-cognitiveservices-speech
pydub